Stop and turn when sensing front wall

+++ PID/pidConstants.h
@@ -28,5 +28,5 @@ const static int LEFT_IR_WALL = 100;
 const static int RIGHT_IR_WALL = 350;
 
 
-const static int DEFAULT_FORWARD_SPEED = 500/1885.0;
+const static double DEFAULT_FORWARD_SPEED = 500/1885.0;
 #endif
